+++ packages/jest-transform/src/types.ts
@@ -15,6 +15,7 @@ export interface ShouldInstrumentOptions
   > {
   changedFiles?: Set<string>;
   sourcesRelatedToTestsInChangedFiles?: Set<string>;
+  globalRootDir?: string;
 }
 
 export interface Options