+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `[jest-reporters]` Fix issue where console output not displayed for GHA reporter even with `silent: false` option ([#15864](https://github.com/jestjs/jest/pull/15864))
 - `[jest-runtime]` Fix issue where user cannot utilize dynamic import despite specifying `--experimental-vm-modules` Node option ([#15842](https://github.com/jestjs/jest/pull/15842))
 - `[jest-test-sequencer]` Fix issue where failed tests due to compilation errors not getting re-executed even with `--onlyFailures` CLI option ([#15851](https://github.com/jestjs/jest/pull/15851))
+- `[jest-runner, jest-transform]` Fix coverage report doesn't show correct code coverage when using `projects` config option ([#15880](https://github.com/jestjs/jest/pull/15880)), fixes ([#5417](https://github.com/jestjs/jest/issues/5417))
 
 ### Chore & Maintenance